+++ crates/swc_ecma_transforms_base/tests/resolver/minifier/7/output.js
@@ -1,0 +1,11 @@
+var a__1 = 1;
+function f__1() {
+    a__2++;
+    try {
+        x();
+    } catch (a__2) {
+        var a__2 = 2;
+    }
+}
+f__1();
+console.log(a__1);