+++ crates/swc_ecma_transforms_base/tests/resolver/minifier/8/input.js
@@ -1,0 +1,11 @@
+var a = 1;
+function f() {
+    a++;
+    try {
+        console;
+    } catch (a) {
+        var a = 2 + x();
+    }
+}
+f();
+console.log(a);