+++ crates/swc_ecma_transforms_base/tests/resolver/minifier/9/output.js
@@ -1,0 +1,15 @@
+var o__1 = 0;
+function f__1() {
+    try {
+        throw 1;
+    } catch (c__3) {
+        try {
+            throw 2;
+        } catch (o__2) {
+            var o__2 = 3;
+            console.log(o__2);
+        }
+    }
+    console.log(o__2);
+}
+f__1();